Added missing functionality

+++ src/main/java/com/bdg/bdgebay/repository/user/UserRepository.java
@@ -8,7 +8,38 @@ import org.springframework.data.repository.CrudRepository;
  */
 
 
-public interface UserRepository extends CrudRepository<User, Integer> {
+public interface UserRepository extends CrudRepository<User, Long> {
 
     User findByIdAndDeletedIsNull(Long userId);
 }
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
